Refactor storage classes.
