+++ src/Entity/Storage/Sql/ContentEntityStorage.php
@@ -10,55 +10,4 @@ class ContentEntityStorage extends SqlContentEntityStorage implements ContentEnt
 
   use ContentEntityStorageTrait;
 
-  /**
-   * {@inheritdoc}
-   */
-  public function getQueryServiceName() {
-    return 'entity.query.sql.multiversion';
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function buildQuery($ids, $revision_id = FALSE) {
-    $query = parent::buildQuery($ids, $revision_id);
-
-    // Prevent to modify the query before entity type updates.
-    if ($ids === NULL) {
-      return $query;
-    }
-
-    $revision_alias = 'revision';
-    if ($this->entityType->isTranslatable()) {
-      // Join the revision data table in order to set the delete condition.
-      $revision_table = $this->getRevisionDataTable();
-      $revision_alias = 'revision_data';
-      if ($revision_id) {
-        $query->join($revision_table, $revision_alias, "$revision_alias.{$this->revisionKey} = revision.{$this->revisionKey} AND $revision_alias.{$this->revisionKey} = :revisionId", array(':revisionId' => $revision_id));
-      }
-      else {
-        $query->join($revision_table, $revision_alias, "$revision_alias.{$this->revisionKey} = revision.{$this->revisionKey}");
-      }
-    }
-    $query->condition("$revision_alias._deleted", (int) $this->isDeleted);
-    // Entities in other workspaces than the active one can only be queried with
-    // the Entity Query API and not by the storage handler itself.
-    $query->condition("$revision_alias.workspace", $this->getActiveWorkspaceId());
-    return $query;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function onTransactionCommit(array $revision_ids) {
-    foreach ($revision_ids as $id => $revision) {
-      $this->database->update($this->baseTable)
-        ->fields(array(
-            $this->revisionKey => $revision,
-          ))
-        ->condition($this->idKey, $id)
-        ->execute();
-    }
-  }
-
 }
